+++ examples/lib/stories/effects/opacity_effect_example.dart
@@ -2,7 +2,6 @@ import 'package:flame/components.dart';
 import 'package:flame/effects.dart';
 import 'package:flame/game.dart';
 import 'package:flame/input.dart';
-
 import '../../commons/ember.dart';
 
 class OpacityEffectExample extends FlameGame with TapDetector {
@@ -32,7 +31,7 @@ class OpacityEffectExample extends FlameGame with TapDetector {
         size: Vector2.all(100),
       )..add(
           OpacityEffect.fadeOut(
-            StandardEffectController(
+            EffectController(
               duration: 1.5,
               reverseDuration: 1.5,
               infinite: true,
@@ -46,9 +45,9 @@ class OpacityEffectExample extends FlameGame with TapDetector {
   void onTap() {
     final opacity = sprite.paint.color.opacity;
     if (opacity >= 0.5) {
-      sprite.add(OpacityEffect.fadeOut(StandardEffectController(duration: 1)));
+      sprite.add(OpacityEffect.fadeOut(EffectController(duration: 1)));
     } else {
-      sprite.add(OpacityEffect.fadeIn(StandardEffectController(duration: 1)));
+      sprite.add(OpacityEffect.fadeIn(EffectController(duration: 1)));
     }
   }
 }