+++ examples/lib/stories/effects/rotate_effect_example.dart
@@ -25,7 +25,7 @@ class RotateEffectExample extends FlameGame {
     compass.rim.add(
       RotateEffect.by(
         1.0,
-        StandardEffectController(
+        EffectController(
           duration: 6,
           reverseDuration: 3,
           curve: Curves.ease,
@@ -37,7 +37,7 @@ class RotateEffectExample extends FlameGame {
       ..add(
         RotateEffect.to(
           Transform2D.tau,
-          StandardEffectController(
+          EffectController(
             duration: 20,
             infinite: true,
           ),
@@ -46,7 +46,7 @@ class RotateEffectExample extends FlameGame {
       ..add(
         RotateEffect.by(
           Transform2D.tau * 0.015,
-          StandardEffectController(
+          EffectController(
             duration: 0.1,
             reverseDuration: 0.1,
             infinite: true,
@@ -56,7 +56,7 @@ class RotateEffectExample extends FlameGame {
       ..add(
         RotateEffect.by(
           Transform2D.tau * 0.021,
-          StandardEffectController(
+          EffectController(
             duration: 0.13,
             reverseDuration: 0.13,
             infinite: true,