+++ packages/flame/lib/src/effects/transform2d_effect.dart
@@ -1,7 +1,7 @@
 import '../components/position_component.dart';
 import '../game/transform2d.dart';
 import 'component_effect.dart';
-import 'effect_controller.dart';
+import 'controllers/effect_controller.dart';
 
 /// Base class for effects that target a [Transform2D] property.
 ///