+++ packages/flame/test/effects/transform2d_effect_test.dart
@@ -1,6 +1,5 @@
 import 'package:flame/src/components/position_component.dart';
-import 'package:flame/src/effects/effect_controller.dart';
-import 'package:flame/src/effects/standard_effect_controller.dart';
+import 'package:flame/src/effects/controllers/effect_controller.dart';
 import 'package:flame/src/effects/transform2d_effect.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -18,12 +17,12 @@ void main() {
         game.add(component);
         game.update(0);
 
-        final effect = _MyEffect(StandardEffectController(duration: 1));
+        final effect = _MyEffect(EffectController(duration: 1));
         component.add(effect);
         game.update(0);
         expect(effect.transform, component.transform);
 
-        final effect2 = _MyEffect(StandardEffectController(duration: 1));
+        final effect2 = _MyEffect(EffectController(duration: 1));
         expect(
           () async => game.add(effect2),
           throwsA(isA<UnsupportedError>()),