+++ packages/flame_test/lib/flame_test.dart
@@ -1,5 +1,6 @@
 export 'src/expect_double.dart';
 export 'src/expect_vector2.dart';
+export 'src/fails_assert.dart';
 export 'src/flame_test.dart';
 export 'src/mock_gesture_events.dart';
 export 'src/mock_image.dart';